+++ CHANGELOG.md
@@ -4,6 +4,16 @@ As of version 2.0.0, all notable changes to HTML Minifier Next (HMN) are documen
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.3.0] - 2026-08-27
+
+### Added
+
+* Added warnings for options that rely on another option, which used to pass silently
+
+### Changed
+
+* Documented which options rely on another option to take effect
+
 ## [8.2.2] - 2026-08-26
 
 ### Changed